refactor(app): S12 workspaceStore/Sidebar/ImageUpload 拆分
将 workspaceStore 拆为 setup/local/sync 模块，Sidebar 源列表下沉至 settings，ImageUpload 抽 useImageUploadFlow 并按 panel 拆分 CSS。
