+++ app/src/features/library/image-upload/ImageUploadDropzone.css
@@ -1,0 +1,271 @@
+/* 拖拽选区 */
+.image-upload-dropzone {
+  border: 2px dashed;
+  border-radius: 1rem;
+  padding: 3rem 2rem;
+  text-align: center;
+  transition: all 0.3s cubic-bezier(0.4, 0, 0.2, 1);
+  cursor: pointer;
+  position: relative;
+  overflow: hidden;
+  min-height: 280px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.image-upload-dropzone::before {
+  content: '';
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background: linear-gradient(
+    135deg,
+    rgba(var(--pix-violet-rgb), 0.05) 0%,
+    rgba(147, 51, 234, 0.05) 100%
+  );
+  opacity: 0;
+  transition: opacity 0.3s ease;
+}
+
+.image-upload-dropzone.active::before {
+  opacity: 1;
+}
+
+.image-upload-dropzone.active {
+  border-color: var(--pix-violet);
+  background-color: var(--pix-violet-soft);
+  border-width: 3px;
+  transform: scale(1.02);
+  box-shadow:
+    0 10px 25px -5px rgba(var(--pix-violet-rgb), 0.2),
+    0 4px 6px -2px rgba(var(--pix-violet-rgb), 0.1);
+}
+
+.image-upload-dropzone.inactive {
+  border-color: #d1d5db;
+  background-color: #ffffff;
+}
+
+.image-upload-dropzone.inactive:hover {
+  border-color: var(--pix-violet);
+  background-color: #f8fafc;
+  border-width: 2.5px;
+  box-shadow: 0 4px 12px rgba(var(--pix-violet-rgb), 0.1);
+}
+
+.image-upload-content {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 1.25rem;
+  position: relative;
+  z-index: 1;
+}
+
+.image-upload-icon-container {
+  padding: 1.25rem;
+  border-radius: 50%;
+  transition: all 0.3s cubic-bezier(0.4, 0, 0.2, 1);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.image-upload-icon-container.active {
+  background: linear-gradient(
+    135deg,
+    var(--pix-violet) 0%,
+    var(--pix-violet-deep) 100%
+  );
+  transform: scale(1.1);
+  box-shadow: 0 8px 16px rgba(var(--pix-violet-rgb), 0.3);
+}
+
+.image-upload-icon-container.inactive {
+  background: linear-gradient(135deg, #f1f5f9 0%, #e2e8f0 100%);
+}
+
+.image-upload-icon-container.inactive:hover {
+  background: linear-gradient(135deg, #e0e7ff 0%, #c7d2fe 100%);
+  transform: translateY(-2px);
+}
+
+.image-upload-icon {
+  width: 2.5rem;
+  height: 2.5rem;
+  transition: all 0.3s ease;
+}
+
+.image-upload-icon.active {
+  color: #ffffff;
+  animation: image-upload-bounce 0.6s ease;
+}
+
+.image-upload-icon.inactive {
+  color: #64748b;
+}
+
+@keyframes image-upload-bounce {
+  0%,
+  100% {
+    transform: translateY(0);
+  }
+  50% {
+    transform: translateY(-8px);
+  }
+}
+
+.image-upload-text {
+  font-size: 1.125rem;
+  font-weight: 600;
+  line-height: 1.5;
+  transition: color 0.3s ease;
+}
+
+.image-upload-text.active {
+  color: #1e40af;
+}
+
+.image-upload-text.inactive {
+  color: #1f2937;
+}
+
+.image-upload-description {
+  font-size: 0.875rem;
+  color: #6b7280;
+  margin-top: 0.5rem;
+  line-height: 1.6;
+  max-width: 500px;
+}
+
+.image-upload-crop-hint {
+  color: var(--pix-violet);
+  font-weight: 600;
+  display: inline-block;
+  margin-left: 0.25rem;
+}
+
+.image-upload-formats {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 0.5rem;
+  justify-content: center;
+  margin-top: 1rem;
+}
+
+.image-upload-format-tag {
+  display: inline-flex;
+  align-items: center;
+  padding: 0.375rem 0.75rem;
+  background-color: #f3f4f6;
+  color: #4b5563;
+  border-radius: 0.5rem;
+  font-size: 0.75rem;
+  font-weight: 500;
+  transition: all 0.2s ease;
+}
+
+.image-upload-format-tag:hover {
+  background-color: #e5e7eb;
+  transform: translateY(-1px);
+}
+
+.image-upload-native-actions {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 0.75rem;
+  justify-content: center;
+  margin-top: 1.25rem;
+}
+
+.image-upload-native-button {
+  display: inline-flex;
+  align-items: center;
+  gap: 0.5rem;
+  padding: 0.625rem 1rem;
+  border: 1px solid #d1d5db;
+  border-radius: 0.5rem;
+  background: #ffffff;
+  color: #374151;
+  font-size: 0.875rem;
+  font-weight: 500;
+  cursor: pointer;
+  transition: all 0.2s ease;
+}
+
+.image-upload-native-button:hover {
+  border-color: var(--pix-violet);
+  color: var(--pix-violet-deep);
+  background: var(--pix-violet-soft);
+}
+
+.image-upload-native-icon {
+  width: 1rem;
+  height: 1rem;
+}
+
+@media (max-width: 640px) {
+  .image-upload-dropzone {
+    padding: 2rem 1.5rem;
+    min-height: 240px;
+  }
+
+  .image-upload-icon {
+    width: 2rem;
+    height: 2rem;
+  }
+
+  .image-upload-icon-container {
+    padding: 1rem;
+  }
+
+  .image-upload-text {
+    font-size: 1rem;
+  }
+
+  .image-upload-description {
+    font-size: 0.8125rem;
+    max-width: 100%;
+  }
+
+  .image-upload-formats {
+    margin-top: 0.75rem;
+    gap: 0.375rem;
+  }
+
+  .image-upload-format-tag {
+    font-size: 0.6875rem;
+    padding: 0.25rem 0.5rem;
+  }
+}
+
+@media (max-width: 480px) {
+  .image-upload-dropzone {
+    padding: 1.5rem 1rem;
+    min-height: 200px;
+  }
+
+  .image-upload-content {
+    gap: 1rem;
+  }
+
+  .image-upload-icon {
+    width: 1.75rem;
+    height: 1.75rem;
+  }
+
+  .image-upload-icon-container {
+    padding: 0.875rem;
+  }
+
+  .image-upload-text {
+    font-size: 0.9375rem;
+  }
+
+  .image-upload-description {
+    font-size: 0.75rem;
+  }
+}
